# docs/swagger.yaml
basePath: /
definitions:
  subscriptions-data-service_internal_model.RecordDto:
    properties:
      end_date:
        type: string
      id:
        type: integer
      price:
        type: integer
      service_name:
        type: string
      start_date:
        type: string
      user_id:
        type: string
    type: object
host: 127.0.0.1:8080
info:
  contact: {}
  title: Subscriptions Data Service
paths:
  /records:
    get:
      parameters:
      - description: Количество записей
        in: query
        name: limit
        type: integer
      - description: Смещение
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/subscriptions-data-service_internal_model.RecordDto'
            type: array
        "400":
          description: некорректное число записей
          schema:
            type: string
        "500":
          description: внутренняя ошибка
          schema:
            type: string
      summary: Получить список подписок
      tags:
      - records
    post:
      consumes:
      - application/json
      parameters:
      - description: Данные подписки
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/subscriptions-data-service_internal_model.RecordDto'
      produces:
      - application/json
      responses:
        "201":
          description: Созданная запись
          schema:
            $ref: '#/definitions/subscriptions-data-service_internal_model.RecordDto'
        "400":
          description: некорректный JSON или диапазон дат
          schema:
            type: string
        "500":
          description: внутренняя ошибка
          schema:
            type: string
      summary: Создать подписку
      tags:
      - records
  /records/{id}:
    delete:
      parameters:
      - description: ID записи
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: некорректный id
          schema:
            type: string
        "404":
          description: запись не найдена
          schema:
            type: string
        "500":
          description: внутренняя ошибка
          schema:
            type: string
      summary: Удалить подписку
      tags:
      - records
    get:
      parameters:
      - description: ID записи
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/subscriptions-data-service_internal_model.RecordDto'
        "400":
          description: некорректный id
          schema:
            type: string
        "404":
          description: запись не найдена
          schema:
            type: string
        "500":
          description: внутренняя ошибка
          schema:
            type: string
      summary: Получить подписку по ID
      tags:
      - records
    put:
      consumes:
      - application/json
      parameters:
      - description: ID записи
        in: path
        name: id
        required: true
        type: integer
      - description: Новые данные
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/subscriptions-data-service_internal_model.RecordDto'
      responses:
        "204":
          description: No Content
        "400":
          description: некорректный JSON, id или диапазон дат
          schema:
            type: string
        "404":
          description: запись не найдена
          schema:
            type: string
        "500":
          description: внутренняя ошибка
          schema:
            type: string
      summary: Обновить подписку
      tags:
      - records
  /records/total:
    get:
      parameters:
      - description: UUID пользователя
        in: query
        name: user_id
        type: string
      - description: Название сервиса
        in: query
        name: service_name
        type: string
      - description: Начало периода (MM-YYYY)
        in: query
        name: from
        required: true
        type: string
      - description: Конец периода (MM-YYYY)
        in: query
        name: to
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: integer
            type: object
        "400":
          description: некорректный формат даты или диапазон дат
          schema:
            type: string
        "500":
          description: внутренняя ошибка
          schema:
            type: string
      summary: Получить сумму подписок за период
      tags:
      - records
swagger: "2.0"